Audit fixes

+++ contracts/Migrations.sol
@@ -1,20 +1,20 @@
 pragma solidity >=0.4.21 <0.6.0;
 
 contract Migrations {
-  address public owner;
-  uint public last_completed_migration;
-  constructor() public {
-    owner = msg.sender;
-  }
-  modifier restricted() {
-    if (msg.sender == owner) _;
-  }
-  function setCompleted(uint completed) public restricted {
-    last_completed_migration = completed;
-  }
+    address public owner;
+    uint256 public last_completed_migration;
+    constructor() public {
+        owner = msg.sender;
+    }
+    modifier restricted() {
+        if (msg.sender == owner) _;
+    }
+    function setCompleted(uint256 completed) public restricted {
+        last_completed_migration = completed;
+    }
 
-  function upgrade(address new_address) public restricted {
-    Migrations upgraded = Migrations(new_address);
-    upgraded.setCompleted(last_completed_migration);
-  }
+    function upgrade(address new_address) public restricted {
+        Migrations upgraded = Migrations(new_address);
+        upgraded.setCompleted(last_completed_migration);
+    }
 }
